Move override config resolution out of shopper-agent-utils
app/ssr.js runs under bare babel-node, which has no webpack aliases, so
importing shopper-agent-utils.js pulled in its
'@salesforce/retail-react-app/app/constants' import and the SSR server failed
to start with MODULE_NOT_FOUND, taking the Lighthouse CI run with it.

Move the override resolution into an import-free app/utils/commerce-client-overrides.js,
following the same convention token-bridge.js documents. ssr.js imports it
relatively while the widget component and shopper-agent-utils use the alias, so
the CSP and the widget still resolve the override from one validated value.
